Route packer UI through actions
